Update ClimbSafeFeatureSet6Controller.java
updated controller methods set 6: added JavaDoc and a check for the constraint on Equipment

+++ ca.mcgill.ecse.climbsafe/app/src/main/java/ca/mcgill/ecse/climbsafe/controller/ClimbSafeFeatureSet6Controller.java
@@ -6,28 +6,50 @@ import ca.mcgill.ecse.climbsafe.application.ClimbSafeApplication;
 import ca.mcgill.ecse.climbsafe.model.ClimbSafe;
 import ca.mcgill.ecse.climbsafe.model.Equipment;
 
+
 public class ClimbSafeFeatureSet6Controller {
 
   public static ClimbSafe climbSafe = ClimbSafeApplication.getClimbSafe();
 
+  /**
+   * Corresponding Feature: DeleteEquipment
+   * 
+   * This method deletes an equipment if it not in a bundle. Else, it throws an exception.
+   * 
+   * @author Asma Gandour
+   * @param name the name of the equipment to be deleted
+   * @throws InvalidInputException
+   */
   public static void deleteEquipment(String name) throws InvalidInputException {
+    var error = "";
+    Equipment equipment = findEquipment(name);
 
-    try {
-      Equipment equipment = findEquipment(name);
+    if (equipment != null && equipment.hasBundleItems()) {
+      error = "The piece of equipment is in a bundle and cannot be deleted";
+    }
 
-      if (equipment != null) {
-        equipment.delete();
-      }
+    if (!error.isEmpty()) {
+      throw new InvalidInputException(error.trim());
+    }
 
-    } catch (RuntimeException e) {
 
-      throw new InvalidInputException(e.getMessage());
+    if (equipment != null) {
+      equipment.delete();
     }
   }
 
   // this method does not need to be implemented by a team with five team members
   public static void deleteEquipmentBundle(String name) {}
 
+  /**
+   * Corresponding Feature: ViewAssignment
+   * 
+   * This method iterates over climbSafe's assignments, converts each one of them into a TO
+   * assignment and returns a list of these TO assignments.
+   * 
+   * @author Asma Gandour
+   * @return the list of TOAssignments for the current climbSafe
+   */
   public static List<TOAssignment> getAssignments() {
 
     var assignments = new ArrayList<TOAssignment>();
@@ -41,14 +63,22 @@ public class ClimbSafeFeatureSet6Controller {
     return assignments;
   }
 
+  /**
+   * 
+   * A helper method for deleteEquipment(). It iterates over climbSafe's list of equipments to find
+   * the equipment with the name corresponding to the input name
+   * 
+   * @author Asma Gandour
+   * @param name the name of the equipment to be found
+   * @return the equipment whose name corresponds to name
+   */
   private static Equipment findEquipment(String name) {
     Equipment foundEquipment = null;
 
     for (Equipment equipment : climbSafe.getEquipment()) {
       if (equipment.getName().equals(name)) {
 
-        foundEquipment =
-            new Equipment(name, equipment.getPricePerWeek(), equipment.getWeight(), climbSafe);
+        foundEquipment = equipment;
 
         break;
       }
